+++ testapps/GHIssues/src/main/java/ghissues/gh1093/Gh1093Department.java
@@ -1,0 +1,9 @@
+package ghissues.gh1093;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Gh1093Department {
+    public String name;
+    public List<Gh1093User> users = new ArrayList<>();
+}